Add files via upload
You can create a personal portfolio with the basic skills in CSS and HTML This is yet again, one of the most comman projects on web development for beginners. In a personal portfolio, you get to show a wide range of samples and skits and add your resume. You can add an image at the top where you can introduce yourself, below that, you can adid contact information. For Demo